fix types

+++ src/touch/xpt2046.c
@@ -119,7 +119,7 @@ void
 xpt2406_interrupt_handler(void * ctx)
 {
 	const spitarget_t target = targettsc1;
-	unsigned x, y, z1, z2;
+	uint_fast16_t x, y, z1, z2;
 	xpt2046_read4(target, & x, & y, & z1, & z2);
 	PRINTF("xpt2046 interrupt: x=%5u, y=%5u z1=%5u, z2=%5u\n", x, y, z1, z2);
 }
@@ -129,13 +129,13 @@ void xpt2046_initialize(void)
 	const spitarget_t target = targettsc1;
 	//PRINTF("xpt2046_initialize:\n");
 	{
-		unsigned x, y, z1, z2;
+		uint_fast16_t x, y, z1, z2;
 		xpt2046_read4(target, & x, & y, & z1, & z2);
 	}
 #if 0
 	for (;;)
 	{
-		unsigned x, y, z1, z2;
+		uint_fast16_t x, y, z1, z2;
 		xpt2046_read4(target, & x, & y, & z1, & z2);
 		PRINTF("xpt2046: x=%5u, y=%5u z1=%5u, z2=%5u\n", x, y, z1, z2);
 	}
